fix page require
