+++ packages/serverless-nextjs-component/__tests__/lambda@edge-handler.test.js
@@ -5,6 +5,16 @@ jest.mock("../manifest.json", () => require("./fixtures/manifest.json"), {
   virtual: true
 });
 
+const mockPageRequire = mockPagePath => {
+  jest.mock(
+    `../${mockPagePath}`,
+    () => require(`./fixtures/built-artifact/${mockPagePath}`),
+    {
+      virtual: true
+    }
+  );
+};
+
 describe("Lambda@Edge", () => {
   describe("Routing", () => {
     it("serves optimised page from S3 static-pages folder", async () => {
@@ -55,6 +65,29 @@ describe("Lambda@Edge", () => {
         }
       });
       expect(request.uri).toEqual("/manifest.json");
+    });
+  });
+
+  describe("When SSR@Edge is enabled", () => {
+    it("renders page at the edge", async () => {
+      const event = createCloudFrontEvent({
+        uri: "/customers",
+        host: "mydistribution.cloudfront.net",
+        origin: {
+          ssr: {
+            domainName: "ssr-api.execute-api.us-east-1.amazonaws.com"
+          }
+        }
+      });
+
+      mockPageRequire("pages/customers/index.js");
+
+      const response = await handler(event, {});
+
+      const decodedBody = new Buffer(response.body, "base64").toString("utf8");
+
+      expect(decodedBody).toEqual("pages/customers/index.js");
+      expect(response.status).toEqual(200);
     });
   });
 });
