+++ packages/serverless-nextjs-component/package.json
@@ -22,7 +22,7 @@
     "@serverless/aws-api-gateway": "^1.0.1",
     "@serverless/aws-cloudfront": "git://github.com/serverless-components/aws-cloudfront.git",
     "@serverless/aws-lambda": "^1.2.0",
-    "@serverless/aws-s3": "^3.0.1",
+    "@serverless/aws-s3": "git://github.com/serverless-components/aws-s3.git",
     "@serverless/backend": "^3.3.0",
     "@serverless/core": "^1.0.0",
     "fs-extra": "^8.1.0",
